introduce new vertx package
moved @Sanne's Vertx instance management to a new
package defined as an extension point
