--- hibernate-reactive-core/src/main/java/org/hibernate/reactive/service/VertxService.java
@@ -1,16 +1,0 @@
-package org.hibernate.reactive.service;
-
-import io.vertx.core.Vertx;
-import org.hibernate.service.Service;
-
-/**
- * Used to get the reference to the Vertx instance to use.
- * This is a service so to allow injecting an external instance,
- * or allow Hibernate Reactive to manage its own instance.
- */
-@FunctionalInterface
-public interface VertxService extends Service {
-
-    Vertx getVertx();
-
-}
